Updated QT_LIB_INCLUDE to use only basic libraries
Updated the comment and contents of QT_LIB_INCLUDE to just use the
essential libraries listed in the .pro file. Since the script already
handles sub-dependencies, the dependencies do not need to be defined
excplicitly.

(Hypothetically, this could be removed altogether... some more testing
is needed to ensure that removing it will still work!)