Enable splitting files in BuildQt56StaticDevArchive.py
Artifactory has a limit of 100 MBs, so file splitting is needed to
keep within that limit. Support has been added to enforce the file
splitting, and to rename the file back to non-splitted mode if only one
part is detected. (For instance, if ARC.7z.001 is the only file found,
then ARC.7z.001 will automatically be renamed to ARC.7z. However, if
ARC.7z.001 and ARC.7z.002 exist, no renaming will take place.)